% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotDiagnosticPlots.R
\name{ggplotDiagnosticPlots}
\alias{ggplotDiagnosticPlots}
\title{ggplotDiagnosticPlots: function for diagnostic plots for INLA model}
\usage{
ggplotDiagnosticPlots(model, observed = NULL, CI = FALSE,
  binwidth = NULL, se = TRUE, method = "auto")
}
\arguments{
\item{model}{Model object from incidenceMapR::modelTrainR}

\item{observed}{The observed values. If NULL defined by model family}

\item{CI}{Add credible intervals to the fitted values?}

\item{binwidth}{The size of the bins used for the histogram. If NULL ggplot guesses for you.}

\item{se}{Plot a ribbon showing the standard error of the smoother.}

\item{method}{What method should be used for the smoother. Defaults to loess unless data is large. Other options include 'gam', 'loess', 'lm'. See geom_smooth for details.}
}
\value{
List of 3 ggplot objects: default plots, observed vs fitted and histogram, standardized residuals
}
\description{
ggplotDiagnosticPlots: function for diagnostic plots for INLA model
}
\examples{
ggplotDiagnosticPlots(model)

}
