% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWrappers.R
\name{ggplotFixedEffects}
\alias{ggplotFixedEffects}
\title{ggplotFixedEffects: function for plotting data, smoothed model, and fixed effects model next to each other}
\usage{
ggplotFixedEffects(model_effects, shp, censusdb = NULL, title = "",
  shape_level = "residence_census_tract")
}
\arguments{
\item{plotDatEffects}{data.frame that joins dbViewR::modeledData}
}
\value{
ggplot object
}
\description{
ggplotFixedEffects: function for plotting data, smoothed model, and fixed effects model next to each other
}
\examples{


}
