% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotWrappers.R
\name{ggplotSmoothMap}
\alias{ggplotSmoothMap}
\title{ggplotSmoothMap: function for plotting data and smoothed model next to each other on map}
\usage{
ggplotSmoothMap(plotDat, shp)
}
\arguments{
\item{plotDat}{data.frame that joins dbViewR::modeledData and shp}

\item{shp}{sf object with GEOID shapes (all higher levels assume iid and not local smoothing)}
}
\value{
ggplot object
}
\description{
ggplotSmoothMap: function for plotting data and smoothed model next to each other on map
}
\examples{
plotDat <- right_join(model$modeledData,shp, by=c('GEOID'))
plotDat$positive[plotDat$positive==0]<-NaN
ggplotSmoothMap(plotDat,shp)

}
